+++ components/MyDateCard/components/DateMap/DateMap.jsx
@@ -1,7 +1,30 @@
 import React from 'react';
+import { Typography, Box } from '@material-ui/core';
 import Constants from '../../../../constants';
 
 const DateMap = ({ placeIds }) => {
+  if (!placeIds || placeIds.length < 2 || !placeIds[0] || !placeIds[0]) {
+    return (
+      <Box
+        display="flex"
+        alignItems="center"
+        padding="20px"
+        width="100%"
+        height="100%"
+        justifyContent="space-around"
+        flexDirection="column"
+      >
+        <Typography align="center" variant="h6">
+          <span role="img" aria-label="House emoji" style={{ fontSize: '42px' }}>
+            🏠
+          </span>{' '}
+          <br />
+          No map needed - enjoy this date at home!
+        </Typography>
+      </Box>
+    );
+  }
+
   let mapSrc = 'https://www.google.com/maps/embed/v1/directions?';
   mapSrc += `origin=place_id:${placeIds[0]}&`;
   mapSrc += `destination=place_id:${placeIds[1]}&`;