+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased](https://github.com/nathanab/beacon/compare/master...HEAD)
+- Add support for `At Home` dates
 
 ## [v2.0.0](https://github.com/nathanab/beacon/compare/v1.2.0...v2.0.0)
 - Refactor app to use server-side rendering via next.js